Feat/secure audit logs (#121)
* feat(m2m): add m2m client management

* feat(students): optimize student record queries

* refactor(slips): improve slip management and ocr

* feat(appointments): refine scheduling and note flow

* chore(bootstrap): configure api main and services

* refactor: purge unreachable dead code

* fix(locations): include SubMun type in city queries

* fix(ocr): categorize client validation failures

* fix(audit): strip sensitive PII from log metadata

* refactor(db): migrate database queries to views

* fix(students): prevent profile nil dereference

+++ internal/features/analytics/models.go
@@ -4,11 +4,11 @@ import "time"
 
 // DemographicStat represents demographic statistics for analytics.
 type DemographicStat struct {
-	Category    string `db:"category"     json:"category"`
-	MaleCount   int    `db:"male_count"   json:"maleCount"`
-	FemaleCount int    `db:"female_count" json:"femaleCount"`
-	Total       int    `db:"total"        json:"total"`
-	RankPos     int    `db:"rank_pos"     json:"rankPos"`
+	Category    string  `db:"category"     json:"category"`
+	MaleCount   int     `db:"male_count"   json:"maleCount"`
+	FemaleCount int     `db:"female_count" json:"femaleCount"`
+	Total       int     `db:"total"        json:"total"`
+	RankPos     float64 `db:"rank_pos"     json:"rankPos"`
 }
 
 // StudentPersonalInfo represents personal information data for analytics.
